feat: add type move table
